42 - Подготовка к переключению на redux
